Add Retrievers (#16)
EMVB introduced methods using intrinsics and faiss' PQ encoder. In order
to fully support this, we have to have the ability to use different
retrievers in the index.

With this PR, IndexIVF no longer has the base logic to search an index
-- this is given to the Retriever. IndexIVF is responsible for managing
the coordination between the database, the encoder, and the retriever.

There are some additional tweaks to build using MKL on linux. This was
added due to unknown slowness seen on my server. After a ton of
debugging, this is likely due to launching threads in openMP and
threading conflicts on my machine.

Supports: #12