Enable passing Python dictionaries in index.add (#28)
This PR changes the external interface for indexing. Collections
translates Python maps to input passages, but index didn't.

We update that so we can pass dictionaries with either `text` or
`embeddings` keys.

There are a few internal interface changes as well. Notably, add_batch
to collection passed by value, and that's been fixed.